New flow `auto_compute_rollout` to automatically generate a rollout plan.
This flow generates a rollout plan according to https://dfinity.atlassian.net/browse/REL-2733
Currently, the rollout plan is logged to the console, which makes it
accessible from the log screen of the last task.

This can soon be either integrated in the standard rollout flow, or
made to start a standard rollout flow.