Upgrade unassigned nodes on Airflow.
This adds a new task to the rollout, that upgrades the unassigned
nodes right after successful upgrade of the NNS (technically, right
after everything else in the flow has been completed).

This new task requires DRE to be downloaded and executed, so that
is exactly what the code does.  DRE is checked every week for updates
and updated accordingly.

This PR may require a restart of the scheduler to roll out completely.