Justify batches.
It was annoying that every batch had a fixed size which did not correspond
to the right edge of the container of the page.  With the additional property
`flex-grow: 1`, batches grow proportionally (in the horizontal axis) so that
they nicely fill the container, irrespective of the browser width.

I will release this on Monday.