Rollout dashboard fixes.
**Fix 1**

When rollout dashboard asks for a rollout's task instances, rollout dashboard
attempts to retrieve the tasks in a paged way to avoid overcoming Airflow's
baked-in 100-item limit.  For some reason, when tasks are retrieved, the
result is unordered, so when page 2, 3, 4 of task instances are fetched, some
task instances are missing and some task instances are duplicated.
https://github.com/apache/airflow/issues/41283 explains the issue in detail.
The impact to us is that, sometimes, subnets have no associated state, despite
the fact that Airflow clearly has progressed on that subnet.

To mitigate this, we transgress the limitation and ask for 500 task instances.
This seems to work fine, much to my surprise!  With that, the issue of rollout
subnets sometimes not having any associated state is fixed.

**Fix 2**

The URL is always present in the rollout data structure, so stop pretending
it is not, and remove unnecessary conditional.

**Fix 3**

Subnet state icons gain a hyperlink to the pertinent running, failed, or last
completed Airflow task instance (specifically, to the log tab).  This way, one
can directly access the log by clicking on the subnet state icon.

This improves usability of the dashboard and speed to diagnose an issue that
may be occurring during a rollout.