Increase efficiency.
Previously, we were sending the full rollout to the frontend.

This is entirely unnecessary unless tasks of the rollout have changed in status.

We now track and use this information to send no data to the frontend (except
for the every-5-second keepalives) when there are no changes to the data
harvested by the backend.