Fix changed logic causing endless spinner on clients.
Issue introduced in https://github.com/dfinity/dre-airflow/pull/24 .
Due to the way testing happens locally, we did not catch this.

When no one is connected via SSE, the message queue used to
broadcast the results of the data update drops all messages when
nobody is connected.  The logic then finds nothing in the queue
when someone connects via SSE, and sends them a 204 no content
message, which causes the spinner to appear.  Since there are
generally no updates after the first request, no updates are
sent to the client.

This is exactly what did not happen during local testing.  During
local testing, someone (me) is always connected via SSE, because
the browser reconnects periodically and, when the daemon starts,
the browser is connected, there is at least one SSE channel,
there is at least one channel receiver, and therefore there
is always data (or rather there will be after a few seconds)
when the browser reconnects.

The fix is simple: when a client connects, the first message it
must receive is the latest data from cache, then as changes appear
in subsequent requests, send updates to the browser.

This was also never caught before simply because we were not
using any changed logic.