Task link fixed.
Whereas before we were updating the subnet link to the task when
we updated the state, now we update the subnet link to point to the
task logs when the actual state of the task corresponds to the
expected state after updating.  Since tasks are processed in
topological order, the latest "active" task wins the link.

What this does in effect is, e.g., if the rollout is waiting for
a vote (`wait_for_revision_election` is in state `Deferred`), now
we link to the task `wait_for_revision_election`, instead of its
predecessor (in state `Complete`) task `create_proposal_if_none_exists`.
This makes sense because the running task (`Deferred` is a form of
running) is no longer the predecessor of `create_proposal_if_none_exists`.

Will roll out later today.