Stop using start_date / logical_date and use data_interval_end to make decisions.
In Airflow, data_interval_end is the DAG run value that represents either the date a manual DAG run was started or the "end of data interval period" (equivalent to the date an automatic DAG run was started).

While this change is compatible with Airflow 2.x, we are preparing for Airflow 3.x when these variables have already changed to meet these uses.