Composite batch specifications.
Now it is possible to specify HostOS batch node selectors as a tree of joins
or intersects, making it possible to express more complex formulas like
"1 node per subnet `UNION` 1 API boundary node".  This is necessary to
safeguard the health of the API boundary node surface of the IC.

The default rollout plan has been updated to take that into account, and
the first canary batch now upgrades a boundary node too, in order to know
whether a new HostOS has a negative impact on boundary nodes.