Make returning to the homescreen fast (#178)
* Speed up rendering of main view by never closing the socket connection to the server.

In a prior iteration, we made the main state socket connection to the server be closed when a batch detail view was navigated into.  This is silly.  Even the batch detail view uses data from it, and it makes no sense to close the connection just to reopen it again.

In this PR, the prior behavior is restored, but this time, we take care to abort the connection once the user navigates away from the app (this should be done by the browser but it pays to pay attention to detail like that).

* Remove unused import.